requirements.txt
tensorflow
pandas
numpy
matplotlib
seaborn
sklearn
streamlit
Pillow
opencv-python-headless